requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ML/Vector dependencies
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Utility dependencies
tqdm>=4.65.0

# Development dependencies
pytest>=7.4.0
requests>=2.31.0
